+++ CHANGELOG.md
@@ -29,6 +29,9 @@ The following changes have been completed, and will be included in the next tagg
   - added `InvalidFilepathDataError`
 * Filepath
   - added `Filepath`
+    - added `Filepath` `base` option
+    - added `Filepath` `pathApi` option
+    - added `Filepath.basename()`
   - added `MakeFilepathOptions`
   - added `isFilepathData()`
   - added `makeFilepath()`